refractor parsing.
