+++ visearch-android/src/main/java/com/visenze/visearch/android/model/ObjectResult.java
@@ -1,9 +1,14 @@
 package com.visenze.visearch.android.model;
 
+import com.google.gson.annotations.SerializedName;
+
 import java.util.List;
 
 public class ObjectResult extends ProductType {
+    @SerializedName("result")
     private List<ImageResult> result;
+
+    @SerializedName("total")
     private Integer total;
 
     public List<ImageResult> getResult() {
