+++ visearch-android/src/main/java/com/visenze/visearch/android/network/APIService.java
@@ -1,6 +1,6 @@
 package com.visenze.visearch.android.network;
 
-import com.visenze.visearch.android.data.ResponseData;
+import com.visenze.visearch.android.model.ResponseData;
 import com.visenze.visearch.android.network.retry.Retry;
 
 import java.util.Map;
@@ -10,7 +10,6 @@ import okhttp3.RequestBody;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.HeaderMap;
-import retrofit2.http.Headers;
 import retrofit2.http.Multipart;
 import retrofit2.http.POST;
 import retrofit2.http.Part;
